+++ v3/pkg/application/application_debug.go
@@ -40,6 +40,11 @@ func newApplication(options Options) *App {
 func (a *App) logStartup() {
 	var args []any
 
+	// BuildInfo is nil when build with garble
+	if BuildInfo == nil {
+		return
+	}
+
 	wailsPackage, _ := lo.Find(BuildInfo.Deps, func(dep *debug.Module) bool {
 		return dep.Path == "github.com/wailsapp/wails/v3"
 	})